Yet another database format, capable of storing CA certs
In the previous iteration, we encountered two issues: Firstly, a *lot*
of certificates are shared. We want to deduplicate them. Additionally,
we might as well store the CA certificates as well (where a much greater
amount of duplicates is expected), since we're already transferring them
in the TLS handshake anyways and right now discard them.